+++ functions/index.js
@@ -2,7 +2,7 @@ const functions = require("firebase-functions")
 const express = require('express')
 const cors = require('cors')
 const { getAllProducts, getProductById, createProduct, updateProduct, deleteProduct, deleteCartItem } = require('./src/products')
-const { deleteCartItem } = require("./src/orders")
+const { deleteCartItem } = require("./src/cart")
 
 const app = express()
 app.use(cors())
@@ -14,5 +14,5 @@ app.patch('/products/:productId', updateProduct)
 app.delete('/products/:productId', deleteProduct)
 
 
-app.delete('/orders/:itemId', deleteCartItem)
+app.delete('/cart/:itemId', deleteCartItem)
 exports.app = functions.https.onRequest(app)
